Surface JS backtraces.
MiniRacer and `source-map-support` already gives a correct backtrace, but it
doesn't surface in actionview. Rails' ExceptionWrapper seems like a good place
to add this support, but the whole debug middleware concept really wants a ruby
backtrace.

We add a custom error with the JS backtrace as the error message. This will show
nicely on the Exception page.