Migrate to jsbundling usage (#2)
This updates Humid to use jsbundling instead of webpacker. Moving forward, we
have to specify the build path of the SSR entry point. This mimics the
position that js-bundling takes where we let the tool build the file at the cost
of hot reloading.

In the process we also move over to github actions.